+++ Source/ANSHuman/ANSHuman/ANSHuman.h
@@ -14,7 +14,7 @@
 
 #include "ANSObject.h"
 
-static const uint8_t ANSHumanChildrenCount = 20;
+static const short ANSHumanChildrenCount = 20;
 
 typedef enum {
     ANSGenderNotDefined,
@@ -26,10 +26,10 @@ typedef struct ANSHuman ANSHuman;
 struct ANSHuman {
     ANSObject _super;
     char *_name;
-    uint8_t _age;
+    short _age;
     ANSGender _gender;
     
-    uint8_t _childrenCount;
+    short _childrenCount;
     
     ANSHuman *_spouse;
     ANSHuman *_mother;
@@ -51,10 +51,10 @@ extern
 char *ANSGetHumanName(ANSHuman *human);
 
 extern
-void ANSSetHumanAge(ANSHuman *human, int _age);
+void ANSSetHumanAge(ANSHuman *human, short _age);
 
 extern
-int ANSGetHumanAge(ANSHuman *human);
+short ANSGetHumanAge(ANSHuman *human);
 
 extern
 void ANSSetHumanGender(ANSHuman *human, ANSGender gender);
@@ -63,7 +63,7 @@ extern
 ANSGender ANSGetHumanGender(ANSHuman *human);
 
 extern
-int ANSGetChildrenCount(ANSHuman *child);
+short ANSGetChildrenCount(ANSHuman *child);
 
 extern
 ANSHuman *ANSGetHumanSpouse(ANSHuman *human);
@@ -75,7 +75,7 @@ extern
 ANSHuman *ANSGetHumanFather(ANSHuman *human);
 
 extern
-ANSHuman *ANSGetChildOfIndex(ANSHuman *human, int index);
+ANSHuman *ANSGetChildOfIndex(ANSHuman *human, short index);
 
 extern
 void ANSHumanGetMarriedWithSpouse(ANSHuman *human, ANSHuman *spouse);
